Fix StatefulOutputPin
According to https://docs.rs/embedded-hal/latest/embedded_hal/digital/v2/trait.StatefulOutputPin.html,
the functions in StatefulOutputPin should not read the actual electrical
value of the pin, but the current state of the output driver.

The fe310 manual states: "Reading the output_val register returns the
written value. Reading the input_val register returns the actual value
of the pin gated by input_en."

So the right thing to do here is to read the corresponding bit in the
output register.